refactor: mome LatestBlockInner into core to be able to expose subscribe method in LatestBlockProvider
